Encrypt messages to devices with Fernet
Previously, I simply used the public RSA key to encrypt our message back
to the device. The upside is this is simpler to work with. But the
downside is the payload size is limited, so if we need to send larger
payloads to the device, we... can't.

This method now involves generating a Fernet key. This Fernet key is
encrypted with our RSA key first. This gives us a fixed 256 byte key. We
then encrypt the actual payload with our generated Fernet key.

This means on the receiving end, they can decrypt the Fernet key with
their private RSA key, then use the decrypted key to decrypt the actual
message.